requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
elastic-transport==8.15.0
elasticsearch==8.15.1
lxml==5.3.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
typing_extensions==4.12.2
urllib3==2.2.3
